Move MemmcachedCacheEntryHttp test files to right directory, other code cleanup.
Includes removing unused parameters, reordering parameters more sensibly, and renaming variables to clearer names.
